requirements.txt
# Web UI framework
streamlit>=1.33.0

# LangChain core and components
langchain==0.2.0
langchain-community==0.2.0
langchain-core==0.2.0
langchainhub==0.1.15

# Hugging Face models (used by LangChain Embeddings)
sentence-transformers>=2.2.2

# FAISS vector store for retrieval
faiss-cpu>=1.7.4

# PDF loading
pypdf>=3.9.1

#other documents loading
python-docx==1.2.0
markdown==3.4.4
unstructured==0.7.5
unstructured[pdf]==0.7.5
beautifulsoup4==4.12.2
lxml==4.9.3
odfpy==1.4.1


# Environment variable handling
python-dotenv>=1.0.1

# Terminal styling
pyfiglet>=1.0.2
termcolor>=2.3.0
rich>=13.7.0

#translator
deep-translator==1.11.4

